requirements_ui.txt
streamlit>=1.35.0
PyPDF2>=3.0.0
python-docx>=1.1.0
